Update FilamentLoadUnloadHandler.cpp
Created a special test version of this handler, to facilitate debugging the nextgen UI.
The message, "Filament Load/Unload complete." is not displaying in the nextgen UI.
It used to display on the previous two screens, as one navigated back from the load/unload screen.  After correcting font size on those two screens (from 10 10 to 10 20), the message stopped appearing at all...
Message length is tested before each M117 message, to select either the Static or the Scrolling Text variable as the target for the message.  In this edit, I have lengthened the message, "Filament Load/Unload..." to read, "Beginning Filament Load/Unload..." and I have shortened the message, "Filament Load/Unload completed." to read, "Load/Unload completed."
IF the M117 Scroller is not displaying any messages, then the Beginning... message will not display and the Load/Unload complete will display.
If it is the final message that is not displaying, I will need another hypothesis...